Backend API/requirements.txt
aniso8601==8.0.0
asgiref==3.4.1
bcrypt==3.2.0
bootstrap4==0.1.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cryptography==35.0.0
DateTime==4.3
Django==3.2.8
django-crispy-forms==1.13.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.26
protobuf==3.18.1
pycparser==2.20
PyJWT==2.3.0
pytz==2020.1
requests==2.26.0
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.4.2
urllib3==1.26.7
Werkzeug==1.0.1
zope.interface==5.4.0
